Update user id type.

+++ src/main/java/com/dghysc/hy/until/SecurityUtil.java
@@ -19,7 +19,7 @@ public class SecurityUtil {
         return (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
     }
 
-    public static BigInteger getUserId() {
+    public static Long getUserId() {
         return ((User) SecurityContextHolder.getContext().getAuthentication().getPrincipal()).getId();
     }
